Moving Amount/AmountFactory to Module per #1153
^^^
