+++ src/modules/temporaryWindowInjection.ts
@@ -20,6 +20,8 @@ import Notifier from './notifications/Notifier';
 import LogBook from './logbook/LogBook';
 import { LogBookTypes } from './logbook/LogBookTypes';
 import ChangelogItems from './changelog/ChangelogItems';
+import Amount from './wallet/Amount';
+import AmountFactory from './wallet/AmountFactory';
 
 Object.assign(<any>window, {
     GameConstants,
@@ -44,4 +46,6 @@ Object.assign(<any>window, {
     LogBook,
     LogBookTypes,
     ChangelogItems,
+    Amount,
+    AmountFactory,
 });
