+++ src/scripts/breeding/Breeding.ts
@@ -1,5 +1,6 @@
 /// <reference path="../../declarations/GameHelper.d.ts" />
 /// <reference path="../../declarations/DataStore/common/Feature.d.ts" />
+///<reference path="../../declarations/wallet/AmountFactory.d.ts"/>
 
 import Currency = GameConstants.Currency;
